add `onCropped` callback
