+++ src/lib/components/ui/image-cropper/types.ts
@@ -4,6 +4,7 @@ import type { HTMLInputAttributes } from 'svelte/elements';
 export type ImageCropperRootProps = WithChildren<{
 	src?: string;
 	open?: boolean;
+	onCropped?: (url: string) => void;
 }>;
 
 export type ImageCropperDialogProps = DialogContentProps;
