+++ src/routes/components/image-cropper/basic.svelte
@@ -1,8 +1,17 @@
 <script lang="ts">
+	import { getFileFromUrl } from '$lib/components/ui/image-cropper';
 	import * as ImageCropper from '$lib/components/ui/image-cropper';
 </script>
 
-<ImageCropper.Root src="https://github.com/shadcn.png">
+<ImageCropper.Root
+	src="https://github.com/shadcn.png"
+	onCropped={async (url) => {
+		// if you need the file for a form you can call getFileFromUrl with the cropped url
+		const file = await getFileFromUrl(url);
+
+		console.log(file);
+	}}
+>
 	<ImageCropper.Trigger />
 	<ImageCropper.Dialog>
 		<ImageCropper.Cropper />
